+++ HISTORY.md
@@ -8,6 +8,9 @@
 
 - Support alternative centroid generation methods in CVTArchive ({pr}`417`,
   {pr}`437`)
+- Add PyCMAEvolutionStrategy for using pycma in ES emitters ({pr}`434`)
+- **Backwards-incompatible:** Add ranking values to evolution strategy tell
+  method ({pr}`438`)
 - **Backwards-incompatible:** Move evolution strategy bounds to init ({pr}`436`)
 - **Backwards-incompatible:** Use seed instead of rng in ranker ({pr}`432`)
 - **Backwards-incompatible:** Replace status and value with add_info ({pr}`430`)
@@ -37,6 +40,7 @@
 
 #### Improvements
 
+- Allow overriding ES in sphere example ({pr}`439`)
 - Use NumPy SeedSequence in emitters ({pr}`431`)
 - Use numbers types when checking arguments ({pr}`419`)
 - Reimplement ArchiveBase using ArrayStore ({pr}`399`)